wrapper: Don't use [7:0] pins
As those are used by the SPI. Which means we will need
to shift up to the 37:8 and emit only 30 bits of numbers.

Also truncate the test-case (really the right way would be
to compute when we hit the high bits and then change the
prev_values and p_prev_values ... but this is OK for right
now).

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>