test: Set all of the wishbone values to zero at init.
The original issue (which Matt fixed) was that wb_active
was depending on wb_cyc_i and wb_stb_i. Since those were
undefined all the logic around wb_active and the decisions
on that ended up being undefined.

That meant the simulator would decide that since it is undefined,
then_ALL_ paths could be taken and simulate every read/write
case statement, causing a ripling effect in the simulator
and leading to the testcase failing.

While in reality if the GDS was used there would be no hardware
failure as the firmware would have reset the wb_cyc_i and wb_stb_i
to known values.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>