verilog/tests: Add irq_user back
To enable it, the wrapper has to poke:

 - 0x2600_0008 = 3, aka reg_mprj_irq = 3

And then if any of the irq[2:0] are high, from the mgmt_soc.v we
should in theory start at:
 parameter [31:0] PROGADDR_IRQ   = 32'h 0000_0000;

(which is a bit of complication as the firmware starts at that
location too, which means we would need to re-patch that code after
the firmware has started).

And looking at the mgmt_soc.v there is a total of 14 lines:

350     always @* begin
351         irq = 0;
352         irq[3] = irq_stall;
353         irq[4] = irq_uart;
354         irq[6] = irq_spi;
355         irq[7] = irq_7;
356         irq[8] = irq_8;
357         irq[9] = irq_spi_master;
358         irq[10] = irq_counter_timer0;
359         irq[11] = irq_counter_timer1;
360         irq[14:12] = user_irq;
361     end
362

And the picrv32.v implies:
1503      ENABLE_IRQ && irq_state[0]: ..  begin
1504              current_pc = PROGADDR_I
1505              irq_active <= 1;
1506              mem_do_rinst <= 1;

.. which implies that irq_state[0] would get be set the moment irq[14:3]
is high, but in fact:

1535     if (ENABLE_IRQ && ((decoder_trigger && !irq_a_pending & ~irq_mask)) || irq_state
1536             irq_state <=
1537                     irq_state == 2'b00 ? 2'b01 :
1538                     irq_state == 2'b01 ? 2'b10 : 2'b00;

Which is so just changing the states.. and there is no reading of 'irq'
wire at all.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>